sw: don't repaint all text frames on selecting and deleting several bullets
This is similar to disabling bullets for a paragraph, but in this case
the text nodes are deleted as well.

It seems to me that both invalidations are only useful in the numberging
(e.g. Arabic) case, and not in the bullet case.

(cherry picked from commit e267fe1a3dffc39bf5076dd6413ed3b0e698378c)

Change-Id: Ie03e2847e5e50d9464399f4ea0840910e43a663a