sw lok: disable field shadings
There are two problems with it:

First, the field shading is rendered behind bullet portions depending on
the cursor position (if the cursor is before the bullet, then we
render it; but not if the cursor is after the bullet), and tiled
rendering assumes that the render result is independent from the cursor
position.

Second, field shading is always painted directly (not going via
invalidation), so it's possible to end up with one tile showing field
shadings and another not, which is inconsistent. And even if it would be
there consistently, that would mean a lot of unnecessary invalidations.

The alternative would be to disable this from online.git's
loolconfig.xcu, but the item's path would end with something like
"ColorScheme['LibreOfficeDev']/WriterFieldShadings", which means it's
not easy to do this in a way that's independent from the product name.

(cherry picked from commit 876da94619d561e4b4c6c7d12a8d6726eb857df8)

Change-Id: I23c62acc3c15966cfba98fcee99cc96854ba4394