Stopgap workaround for misspelling wavy underline drawing bug on iOS
Explicity make the rectangle white before drawing the red wavy line.
Sure, this means that it looks bad if the document background colour
is not white. Better than how it looks currently, though, with the
rectangle containing the wavy underline ending up (for unknown
reasons) black. And most documents have a white background.

Change-Id: I92b2246bbdbd1404e4c48292c3610d1582a56163
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121370
Tested-by: Tor Lillqvist <tml@collabora.com>
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122906
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>