svx: improve SdrObjEditView::DisposeUndoManager()
Addresses the comment at
<https://gerrit.libreoffice.org/c/core/+/122151/1/svx/source/svdraw/svdedxv.cxx#2765>,
i.e. there can be the case when the undo manager is not an
sw::UndoManager (or any other high level undo manager), but it's an
EditUndoManager, and in that case we should not clear away the owning
pointer to avoid a memory leak.

(cherry picked from commit 4cf51357b75c254e7962dd86140c69474fff129f)

Change-Id: I016cfad0563814d65c8523999c82be7747377171
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/122509
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>