sc: EMBED_SOURCE: extend copy to clip area...
to include the draw objects in the sheet read from system clipboard.

This is needed because GetCellArea() computes the "last row" and "last
column" only based on the cell contents and not the drawing layer
contents.  So copying a "empty cells" range that has images in it and
pasting it will not include any images without this fix.

Change-Id: I56ffb14e881a1aecc5b43590ea4fb3fc1c35ace8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/124629
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Dennis Francis <dennis.francis@collabora.com>