avoid only painting to windows in LOK mode, not invalidations
9b73d3e8926d5f9b10464d19b539eef3eb088f50 disabled painting
to windows in LOK mode, because it doesn't make sense as
those windows are not even visible, but it tried to do it
by avoiding even invalidations. That can trigger problems
because some dialogs actually are forwarded to clients in LOK
mode, and apparently they need the invalidations to proceed
normally (the Format->Cell dialog in Online didn't repaint
scrolling of the Date listbox, I don't know what exactly
the actual problem was, but mnPaintFlags were different
than when not avoiding invalidations).

So revert most of the relevant commits and keep only avoiding
actual drawing, which is what 057968bbce406efe6564347329df45b7e
did.

Change-Id: I4e212966830f4c690e6143b9d5fedbc24619e1ed
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/125673
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Jan Holesovsky <kendy@collabora.com>