do not unnecessarily lock XLayoutManager for menubar changes
The XLayoutManager documentation says that lock() prohibits layout
updates until unlock() is called, so it's not any kind of mutex
or anything, it just allows grouping updates. Which means it's
useless to put around a single call to change menubar visibility.
To make things worse, unlock() will always trigger a layout,
so even if nothing changes, it still does an unnecessary layout
update. And since this gets called from SfxLokHelper::setView(),
it may get called often and it's rather costly.

Change-Id: I3e3e37b6763b15d454e76b4bef0b22f84cf479d5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/126445
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>