vba: fix registering shortcuts keys defined by the vba macros
On issue with registering was that the registering happened when
the macro source was in the process to be read into the library,
which is just a bit too early, because the macro wasn't found and
not registered.
Another issue was with searching for the macro method (hasMacro),
which doesn't search the same when the module name is known and
when it isn't. This was changed so we just iterate through the
modules and call the same "FindMethod" method without any extra
restrictions.

Change-Id: I2f4f2f6d8186b289867456ebdccad27ce8eee231
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129197
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>