lok: no need to layout comments in writer
Annotations in writer use EditEngine what caused
to send cursor position from that EditEngine to
online and showing cursor in the top left corner of
a document after comment insertion.

We don't need to resize/update position comment
windows as we send only abstract description of
the comments and later rendering is handled by client.

Change-Id: I7e3328bb9ba2334e8008b98a6b6edfd53c3fea58