speed up scanning the LOK queue
we frequently scan the queue to caolesce events.
Most of the time we are scanning based on the event type.
So split the queue data into a compact queue that only contains the
type, and another queue for the rest of the data.
That makes the scanning __much__ more cache-friendly.

Change-Id: I618bae754f438ebcc5b3af35f827f350e0deebb0