SVG export: don't loose tab portions
Impress slides are exported to SVG via a metafile. In case shape text
has a string like "A\tB", that was exported as "AB", loosing the tab
character.

The relevant shape is exported as a metafile, with two text array
metafile actions (one for "A", one for "B") and the tab needs no
explicit action as both "A" and "B" has an explicit position. The SVG
contained no explicit position for "B", so the tab was lost visually.

Fix the problem by detecting when the doc model has a "\t" without a
matching metafile action and emitting the explicit position in that
case.

An alternative would be to use implWriteTextPortion() to write a <tspan>
for the tab, but that doesn't guarantee that the width of the tab in the
SVG is the same as in Impress.

(cherry picked from commit e7ca10a35fd665722c67eea85e15f71c11c66248)

Change-Id: Ie21bffce9d09194159e460fb28ab23fbefd52102
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141026
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Marco Cecchetti <marco.cecchetti@collabora.com>
(cherry picked from commit cd04f3973c33e081fb0feb176e7c5e4f11759d77)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/141065
Reviewed-by: Andras Timar <andras.timar@collabora.com>